一日の始めに今日の予定と天気を教えてくれる機能　の最高気温null修正分 もともと実装してた分をfunctionに移動。     validate_temp = r['forecasts'][0]['temperature']['max'] ここ1回しか使わないので変数にする必要ないですか？ r['forecasts'][0]['temperature']['max']でも動いたのですが、 変数にした方が綺麗に見えるかなーと
